refactor(app): share external result question selectors
